Separate controllers to folders resolves #138
