+++ src/main/java/module-info.java
@@ -8,4 +8,6 @@ module ee.ut {
     opens ee.ut.dataObjects to javafx.base, com.fasterxml.jackson.databind;
     exports ee.ut.dataObjects to com.fasterxml.jackson.databind;
     exports ee.ut;
+    exports ee.ut.controllers;
+    opens ee.ut.controllers to javafx.fxml;
 }
