+++ src/main/resources/ee/ut/settings.fxml
@@ -4,7 +4,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 <?import java.lang.String?>
-<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ee.ut.GeneralSettingsController">
+<VBox alignment="CENTER" spacing="20.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ee.ut.controllers.GeneralSettingsController">
     <children>
         <VBox alignment="TOP_LEFT" spacing="20.0"  style="-fx-padding: 0 5 75 20;">
             <HBox alignment="TOP_LEFT" spacing="10.0">
